changed proxy path and policy names; implemented endpoint encoding for udca; made provision tests work
