+++ cmd/provision/provision.go
@@ -15,12 +15,7 @@
 package provision
 
 import (
-	"bytes"
 	"crypto/tls"
-	"crypto/x509"
-	"encoding/base64"
-	"encoding/json"
-	"encoding/pem"
 	"fmt"
 	"io/ioutil"
 	"net"
@@ -35,7 +30,6 @@ import (
 	"github.com/pkg/errors"
 	"github.com/spf13/cobra"
 	"go.uber.org/multierr"
-	"gopkg.in/yaml.v3"
 )
 
 const (
@@ -52,13 +46,6 @@ const (
 	productsURLFormat     = "%s/products"     // RemoteServiceProxyURL
 	verifyAPIKeyURLFormat = "%s/verifyApiKey" // RemoteServiceProxyURL
 	quotasURLFormat       = "%s/quotas"       // RemoteServiceProxyURL
-
-	fluentdInternalFormat = "apigee-udca-%s-%s.%s:20001" // org, env, namespace
-	defaultApigeeCAFile   = "/opt/apigee/tls/ca.crt"
-	defaultApigeeCertFile = "/opt/apigee/tls/tls.crt"
-	defaultApigeeKeyFile  = "/opt/apigee/tls/tls.key"
-
-	policySecretNameFormat = "%s-%s-policy-secret"
 )
 
 type provision struct {
@@ -254,6 +241,12 @@ func (p *provision) run(printf shared.FormatFn) error {
 
 	verifyErrors := p.verify(config, verbosef)
 
+	if p.IsGCPManaged {
+		if err := p.checkRuntimeVersion(config, verbosef); err != nil {
+			return errors.Wrap(err, "fetching runtime version")
+		}
+	}
+
 	if err := p.printConfig(config, printf, verifyErrors); err != nil {
 		return errors.Wrapf(err, "generating config")
 	}
@@ -323,127 +316,6 @@ func (p *provision) verify(config *server.Config, verbosef shared.FormatFn) erro
 	}
 
 	return verifyErrors
-}
-
-func (p *provision) createConfig(cred *keySecret) *server.Config {
-	config := &server.Config{
-		Tenant: server.TenantConfig{
-			InternalAPI:            p.InternalProxyURL,
-			RemoteServiceAPI:       p.RemoteServiceProxyURL,
-			OrgName:                p.Org,
-			EnvName:                p.Env,
-			AllowUnverifiedSSLCert: p.InsecureSkipVerify,
-		},
-	}
-
-	if cred != nil {
-		config.Tenant.Key = cred.Key
-		config.Tenant.Secret = cred.Secret
-	}
-
-	if p.IsGCPManaged {
-		config.Tenant.InternalAPI = "" // no internal API for GCP
-		config.Analytics.CollectionInterval = 10 * time.Second
-
-		config.Analytics.FluentdEndpoint = fmt.Sprintf(fluentdInternalFormat, p.Org, p.Env, p.Namespace)
-		config.Analytics.TLS.CAFile = defaultApigeeCAFile
-		config.Analytics.TLS.CertFile = defaultApigeeCertFile
-		config.Analytics.TLS.KeyFile = defaultApigeeKeyFile
-	}
-
-	if p.IsOPDK {
-		config.Analytics.LegacyEndpoint = true
-	}
-
-	return config
-}
-
-func (p *provision) printConfig(config *server.Config, printf shared.FormatFn, verifyErrors error) error {
-	// encode config
-	var yamlBuffer bytes.Buffer
-	yamlEncoder := yaml.NewEncoder(&yamlBuffer)
-	yamlEncoder.SetIndent(2)
-	err := yamlEncoder.Encode(config)
-	if err != nil {
-		return err
-	}
-	configYAML := yamlBuffer.String()
-
-	data := map[string]string{"config.yaml": configYAML}
-	configCRD := server.ConfigMapCRD{
-		APIVersion: "v1",
-		Kind:       "ConfigMap",
-		Metadata: server.Metadata{
-			Name:      "apigee-remote-service-envoy",
-			Namespace: p.Namespace,
-		},
-		Data: data,
-	}
-
-	yamlBuffer.Reset()
-	yamlEncoder = yaml.NewEncoder(&yamlBuffer)
-	yamlEncoder.SetIndent(2)
-	err = yamlEncoder.Encode(configCRD)
-	if err != nil {
-		return err
-	}
-
-	// secret for IsGCPManaged
-	if p.IsGCPManaged {
-		privateKeyBytes := pem.EncodeToMemory(&pem.Block{Type: server.PEMKeyType,
-			Bytes: x509.MarshalPKCS1PrivateKey(config.Tenant.PrivateKey)})
-
-		// create CRD for secret
-		jwksBytes, err := json.Marshal(config.Tenant.JWKS)
-		if err != nil {
-			return err
-		}
-
-		props := map[string]string{server.SecretPropsKIDKey: config.Tenant.PrivateKeyID}
-		propsBuf := new(bytes.Buffer)
-		if err := server.WriteProperties(propsBuf, props); err != nil {
-			return err
-		}
-
-		secretData := map[string]string{
-			server.SecretJKWSKey:    base64.StdEncoding.EncodeToString(jwksBytes),
-			server.SecretPrivateKey: base64.StdEncoding.EncodeToString(privateKeyBytes),
-			server.SecretPropsKey:   base64.StdEncoding.EncodeToString(propsBuf.Bytes()),
-		}
-
-		secretCRD := server.SecretCRD{
-			APIVersion: "v1",
-			Kind:       "Secret",
-			Type:       "Opaque",
-			Metadata: server.Metadata{
-				Name:      fmt.Sprintf(policySecretNameFormat, p.Org, p.Env),
-				Namespace: p.Namespace,
-			},
-			Data: secretData,
-		}
-
-		err = yamlEncoder.Encode(secretCRD)
-		if err != nil {
-			return err
-		}
-	}
-
-	platform := "GCP"
-	if p.IsLegacySaaS {
-		platform = "SaaS"
-	}
-	if p.IsOPDK {
-		platform = "OPDK"
-	}
-
-	printf("# Configuration for apigee-remote-service-envoy (platform: %s)", platform)
-	printf("# generated by apigee-remote-service-cli provision on %s", time.Now().Format("2006-01-02 15:04:05"))
-	if verifyErrors != nil {
-		printf("# WARNING: verification of provision failed. May not be valid.")
-	}
-	printf(yamlBuffer.String())
-
-	return nil
 }
 
 // verify GET RemoteServiceProxyURL/certs
