+++ proxies/remote-proxy-gcp/apiproxy/policies/Get-Runtime-Version.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
+<Script timeLimit="20000" async="false" continueOnError="false" enabled="true" name="Get-Runtime-Version">
+    <DisplayName>Get Runtime Version</DisplayName>
+    <Properties/>
+    <ResourceURL>py://get-runtime-version.py</ResourceURL>
+</Script>
