+++ proxies/remote-proxy-gcp/apiproxy/proxies/default.xml
@@ -161,19 +161,19 @@
             </Response>
             <Condition>(proxy.pathsuffix MatchesPath "/quotas") and (request.verb = "POST")</Condition>
         </Flow>
-        <Flow name="Get Hybrid Version">
-            <Description>Get the version of the hybrid runtime</Description>
+        <Flow name="Get Runtime Env">
+            <Description>Get the version of the apigee runtime</Description>
             <Request>
                 <Step>
-                    <Name>Get-Hybrid-Version</Name>
+                    <Name>Get-Runtime-Version</Name>
                 </Step>
             </Request>
             <Response>
                 <Step>
-                    <Name>Send-Hybrid-Version</Name>
+                    <Name>Send-Runtime-Env</Name>
                 </Step>
             </Response>
-            <Condition>(proxy.pathsuffix MatchesPath "/hybrid-version") and (request.verb = "GET")</Condition>
+            <Condition>(proxy.pathsuffix MatchesPath "/runtime-env") and (request.verb = "GET")</Condition>
         </Flow>
         <Flow name="Unknown Request">
             <Description>Unknown Request</Description>
