+++ proxies/remote-proxy-gcp/apiproxy/resources/py/get-runtime-version.py
